third_party/WebKit/LayoutTests/fast/events/move-event-handler-between-framehosts-expected.txt

This test checks that moving an element with an event handler from one window (FrameHost) to another properly migrates the handler too.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.scrollEventHandlerCount(document) is 0
PASS window.internals.scrollEventHandlerCount(document) is 1
PASS successfullyParsed is true

TEST COMPLETE
